collapse_list_to_strings {SDGupdater}	R Documentation

Turn a list of character vectors/matrices into a vector of character strings

Description

Turns a list of vectors/matrices that hold only character strings into a single vector of character strings, separated by ", ". The elements that make up each component of the input list are pasted into a single string. Each element of the new vector is formed from one component of the input list.

Usage

collapse_list_to_strings(input_list)

Arguments

input_list
List of vectors or matrices, or a vector.

Details

(Initially called this function list_into_vector_of_strings)

Value

Each element of a list is collapsed to a string. Each element of the list becomes an element of a vector. If the input_list is a matrix, the order is first down rows then across columns.

Examples

matrix_1 <- as.matrix(c("England", "UK"))
matrix_2 <- as.matrix(c("Wales"))
matrix_3 <- as.matrix(c("NA"))
matrix_list <- list(matrix_1, matrix_2, matrix_3)
collapse_list_to_strings(matrix_list)

[Package SDGupdater version 0.0.0.9000 Index]