get_all_country_names {SDGupdater}	R Documentation

Extract country from string containing country name

Description

Identifies whether a string contains the name of one or more of the UK countries and if it does, it gives the names of that country as a string, otherwise it returns NA.

Usage

get_all_country_names(variable)

multiple_country_warning(filename, tab, description)

Arguments

variable
Character vector you want to check to see if it contains the name of a UK country.

filename
string giving name of the file (so warning is specific)

tab
string giving name of the tab (so warning is specific)

description
string describing the data that is affected (e.g. which disaggregation)

Details

Future improvements - return countries with correct capitalisation (possibly as option in arguments)

Value

Character vector giving name of country or NA.

a warning

Functions

multiple_country_warning: Extract country from string containing country name

Examples

test_dat <- c("England is a country of the UK", "this does not contain a
country name")
get_all_country_names(test_dat)

multiple_country_warning("foo.xlsx", "Tab 1", "sex_by_age disaggregation")

[Package SDGupdater version 0.0.0.9000 Index]