get_characters_after_dot {SDGupdater}	R Documentation

Get all charcters following a dot

Description

Get all charcters following the last period in a character string

Usage

get_characters_after_dot(string)

Arguments

variable
Character string.

Value

Character string. If there was no period, returns a blank ("")

Examples

get_characters_after_dot("ab.cde")
get_characters_after_dot("ab.cde.fgh")

[Package SDGupdater version 0.0.0.9000 Index]