get_info_cells {SDGupdater}	R Documentation

Isolate cells above column headings

Description

Gets year and country info from the rows above the column headings in a dataframe created using xlsx_cells

Usage

get_info_cells(dat, first_header_row)

Arguments

dat
Dataframe imported using xlsx_cells where there are header cells above the column names containing info on date and year.

first_header_row
Numeric value. The row number in the excel file containing the first set of column headings.

Details

Assumes that the cell containing the year ends with the year

Value

Tibble containing all year and country info.

Examples

first_header_row <- 4
test_dat <- data.frame(row = c(1:5), character = c("data for 2017 ", "England and", "Wales", "impala", "elephant"))
get_info_cells(test_dat, first_header_row)

[Package SDGupdater version 0.0.0.9000 Index]