remove_blanks_and_info_cells {SDGupdater}	R Documentation

Remove blanks and cells above the column headings

Description

Remove blanks and cells above the column headings to clean xlsx_cells data. Do this before using behead()

Usage

remove_blanks_and_info_cells(dat, first_header_row)

Arguments

dat
Tibble originaly imported using xlsx_cells.

first_header_row
Numeric. Row number of first row in the xlsx file containing column names.

Value

Tibble.

Examples

test_dat <- dplyr::tibble(row = c(1:6), character = c("2017", NA, "England",
"Rates", NA, NA), is_blank = c(FALSE, TRUE, FALSE, FALSE, TRUE, FALSE),
data_type = c("character", "blank", "character", "character", "blank",
"numeric"), numeric = c(NA, NA, NA, NA, NA, 20))
remove_blanks_and_info_cells(test_dat, 4)

[Package SDGupdater version 0.0.0.9000 Index]