remove_footnotes {SDGupdater}	R Documentation

Remove footnotes below table

Description

Assuming footnotes are in the first few columns, this function removes those rows, leaving only rows with data. Footnotes must be in one or more of the first x columns, where x = check_columns. If the data are likely to have non-footnote (real) data in the first x column(s) and NAs in all other columns DO NOT use this function as it will likely remove more than just footnotes

remove_footnotes starts at the bottom of the dataframe and stops looking for footnotes as soon as a row that looks like a non-footnote row is encountered. Therefore, you must enter the maximum column number in which you expect to find footnotes.

Usage

remove_footnotes(data, check_columns = 2)

Arguments

data
dataframe with footnotes in which the footnote is followed by NA cells on the same row.

check_columns
numeric. The maximum column index in which you expect there to be footnotes. The default is 2.

Value

data frame with footnote and all-NA rows removed. If no footnotes are identified, the original dataframe is returned.

Examples

test_data <- data.frame("A" = c(10, 5, "footnote:"),
                        "B" = c(1, 3, "this is a footnote"),
                        "C" = c(4, 7, NA),
                        "D" = c(5, 9, NA))
remove_footnotes(test_data, 2)
                        

[Package SDGupdater version 0.0.0.9000 Index]