unique_to_string {SDGupdater}	R Documentation

Unique entries to string

Description

Turns a vector that hold only character strings into a single character string, with values separated by separated by ", ". NAs are ignored.

Usage

unique_to_string(variable)

Arguments

variable
Vector of character strings.

Value

Single character string of unique entries, not including NAs.

Examples

Year <- c("2018", "2015, 2016", NA)
Country <- c(NA, NA, "United Kingdom")
unique_to_string(Year)
unique_to_string(Country)

[Package SDGupdater version 0.0.0.9000 Index]